Remove edge_list from p2t::Point
Nonsense, that modified data owned by the user. Indeed, in poly2tri
Points are allocated by the user and passed as pointers to the lib.

The edge_list inside of Point could also create bugs if one polyline
was used as input for several consecutive CDT.

The edge_list is now aggregated with the user's Points in a data
structure private to the SweepContext.