Clean Node deletion and have a single sweepline data structure
The convex hull finalization requires a "back front" sweepline for which
the same data structure as the "advancing front" can be used.

To avoid confusion during the triangulating when mapping triangles
to the front nodes (which front?), we enforce a single sweepline
(Sweep::front_) that is used for the advancing front and the back front.
Prior to creating the back front the advancing front is deleted and
its nodes are released.