About this fork of the poly2tri library
In this fork of the library I share a number of changes I made to
poly2tri for my own purpose. Some of thoses changes impact the API
and are listed in the README file.

There is no intent to provide any level of maintainance or support of
this repository.

The repository is open source and closed to contributions.