Pre-allocated triangles in SweepContext
Knowing the number of points in the input gives us an upper bound of
the number of triangles in the Delaunay triangulation.
Allocate a contiguous buffer with enough capacity to fit that many triangles.
This avoids allocating each Triangle as a unique_ptr.
And this also allows changing the API method GetTriangles to something
closer to the original API.