Add a memory doubling allocation policy
That policy is well suited when the CDT object is used multiple times
to triangulate inputs of varying sizes. In that case the cost of memory
allocations for points, edges and triangles is amortized over time.

A compile flag is present to fall back to the legacy behavior which
was to reserve only the memory necessary given the size of the input.